Adding prorata temporis computation
