+++ openfisca_pf/tests/daf/redevance_domaniale/tests_fonctionnels/test_calcul_duree.yaml
@@ -1,0 +1,171 @@
+- name: 1 - Calcul en milieu du même mois
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2023-02-10
+      date_fin_occupation: 2023-02-15
+  output:
+    duree_comptable_entre_deux_dates: 6
+    duree_calendaire_entre_deux_dates: 6
+- name: 2 - Calcul en début de mois
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2023-02-01
+      date_fin_occupation: 2023-02-05
+  output:
+    duree_comptable_entre_deux_dates: 5
+    duree_calendaire_entre_deux_dates: 5
+- name: 3 - Calcul mensuel d'un mois en 30 jour (Avril)
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2023-04-01
+      date_fin_occupation: 2023-04-30
+  output:
+    duree_comptable_entre_deux_dates: 30
+    duree_calendaire_entre_deux_dates: 30
+- name: 4 - Calcul mensuel d'un mois en 31 jour (Janvier)
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2023-01-01
+      date_fin_occupation: 2023-01-31
+  output:
+    duree_comptable_entre_deux_dates: 30
+    duree_calendaire_entre_deux_dates: 31
+- name: 5 - Calcul mensuel d'un mois en 28 jour (février 2023)
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2023-02-01
+      date_fin_occupation: 2023-02-28
+  output:
+    duree_comptable_entre_deux_dates: 30
+    duree_calendaire_entre_deux_dates: 28
+- name: 6 - Calcul mensuel d'un mois en 29 jour (février 2024)
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2024-02-01
+      date_fin_occupation: 2024-02-29
+  output:
+    duree_comptable_entre_deux_dates: 30
+    duree_calendaire_entre_deux_dates: 29
+- name: 7 - Calcul annuel pour une année non bissextile
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2022-02-25
+      date_fin_occupation: 2023-02-24
+  output:
+    duree_comptable_entre_deux_dates: 360
+    duree_calendaire_entre_deux_dates: 365
+- name: 8 - Calcul annuel pour une année bissextile
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2024-02-25
+      date_fin_occupation: 2025-02-24
+  output:
+    duree_comptable_entre_deux_dates: 360
+    duree_calendaire_entre_deux_dates: 366
+- name: 9 - Calcul entre février et mars de la même année
+  period: 2018-01-01
+  input:
+    personne:
+      nature_emprise_occupation_redevance_domaniale: ex_eco_01_dpf_hors_entite_publique
+      duree_occupation_redevance_domaniale: 5
+      unite_duree_occupation_redevance_domaniale: Annees
+      zone_occupation_redevance_domaniale: zone_1
+      activite_cultuelle: false
+      nombre_unite_redevance_domaniale: 1
+      variable_redevance_domaniale: 1000
+      nombre_participant_redevance_domaniale: 0
+      commune_redevance_domaniale: com350
+      zone_domaine_prive: sainte_amelie_quartier_des_institutions
+      date_debut_occupation: 2023-02-25
+      date_fin_occupation: 2023-03-01
+  output:
+    duree_comptable_entre_deux_dates: 7
+    duree_calendaire_entre_deux_dates: 5
