+++ openfisca_pf/variables/daf/redevance_domaniale/computation_redevance_domaniale.py
@@ -12,84 +12,6 @@ from openfisca_pf.variables.daf.redevance_domaniale.enums.enums import *
 from openfisca_pf.base import *
 
 
-class duree_occupation_redevance_domaniale_annee(Variable):
-    value_type = float
-    entity = Personne
-    definition_period = DAY
-    label = "La durée d'occupation en année"
-
-    def formula(personne, period, parameters):
-        # Variables
-        duree_occupation_redevance_domaniale = personne('duree_occupation_redevance_domaniale', period)
-        unite_duree_occupation_redevance_domaniale = personne('unite_duree_occupation_redevance_domaniale', period)
-        # Constantes
-        nombre_jour_par_an = parameters(period).daf.redevance_domaniale.constantes.nombre_jour_par_an_rd
-        nombre_mois_par_an = parameters(period).daf.redevance_domaniale.constantes.nombre_mois_par_an_rd
-        nombre_heure_par_jour = parameters(period).daf.redevance_domaniale.constantes.nombre_heure_par_jour_rd
-
-        value = select(
-            [unite_duree_occupation_redevance_domaniale == UnitesDuree.Annees,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Mois,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Jours,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Heures],
-            [duree_occupation_redevance_domaniale,
-            duree_occupation_redevance_domaniale / nombre_mois_par_an,
-            duree_occupation_redevance_domaniale / nombre_jour_par_an,
-            duree_occupation_redevance_domaniale / (nombre_jour_par_an * nombre_heure_par_jour)],
-            )
-        return value
-
-
-class duree_occupation_redevance_domaniale_jour(Variable):
-    value_type = float
-    entity = Personne
-    definition_period = DAY
-    label = "La durée d'occupation en jour"
-
-    def formula(personne, period, parameters):
-        # Variables
-        duree_occupation_redevance_domaniale = personne('duree_occupation_redevance_domaniale', period)
-        unite_duree_occupation_redevance_domaniale = personne('unite_duree_occupation_redevance_domaniale', period)
-        # Constantes
-        nombre_jour_par_an = parameters(period).daf.redevance_domaniale.constantes.nombre_jour_par_an_rd
-        nombre_jour_par_mois = parameters(period).daf.redevance_domaniale.constantes.nombre_jour_par_mois_rd
-
-        value = select(
-            [unite_duree_occupation_redevance_domaniale == UnitesDuree.Annees,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Mois,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Jours],
-            [duree_occupation_redevance_domaniale * nombre_jour_par_an,
-            duree_occupation_redevance_domaniale * nombre_jour_par_mois,
-            duree_occupation_redevance_domaniale],
-            )
-        return value
-
-
-class duree_occupation_redevance_domaniale_mois(Variable):
-    value_type = float
-    entity = Personne
-    definition_period = DAY
-    label = "La durée d'occupation en mois"
-
-    def formula(personne, period, parameters):
-        # Variables
-        duree_occupation_redevance_domaniale = personne('duree_occupation_redevance_domaniale', period)
-        unite_duree_occupation_redevance_domaniale = personne('unite_duree_occupation_redevance_domaniale', period)
-        # Constantes
-        nombre_jour_par_an = parameters(period).daf.redevance_domaniale.constantes.nombre_jour_par_an_rd
-        nombre_jour_par_mois = parameters(period).daf.redevance_domaniale.constantes.nombre_jour_par_mois_rd
-
-        value = select(
-            [unite_duree_occupation_redevance_domaniale == UnitesDuree.Annees,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Mois,
-            unite_duree_occupation_redevance_domaniale == UnitesDuree.Jours],
-            [duree_occupation_redevance_domaniale * nombre_jour_par_mois / nombre_jour_par_an,
-            duree_occupation_redevance_domaniale,
-            duree_occupation_redevance_domaniale / nombre_jour_par_mois],
-            )
-        return value
-
-
 class type_calcul_redevance_domaniale(Variable):
     value_type = str
     entity = Personne
